refactor: 함수 이름 변경 __Joshua -> Joshua__
