+++ frontend/src/page/personnel-info/list/KindergartenInfo.js
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
-import PersonnelList from "../../components/PersonnelList";
+import PersonnelList from "../../../components/PersonnelList";
 
-const ListPaulCommunity = () => {
+const KindergartenList = () => {
   // useState
   const [infoList, setInfoList] = useState([]);
 
@@ -26,4 +26,4 @@ const ListPaulCommunity = () => {
   );
 };
 
-export default ListPaulCommunity;
+export default KindergartenList;
